Co-locate save functionality.
